Fix chrome/browser_ui build error on mac 64-bit.
Review URL: https://chromiumcodereview.appspot.com/11363193

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@167191 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 5248c4f4343136035b1b6b5beb659e55720157ce

+++ chrome/browser/ui/cocoa/bookmarks/bookmark_bar_folder_controller.mm
@@ -872,7 +872,7 @@ struct LayoutMetrics {
 - (int)indexOfButton:(BookmarkButton*)button {
   if (button == nil)
     return -1;
-  int index = [buttons_ indexOfObject:button];
+  NSInteger index = [buttons_ indexOfObject:button];
   return (index == NSNotFound) ? -1 : index;
 }
